requirements.txt
contourpy==1.3.1
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
fonttools==4.55.3
imageio==2.36.1
joblib==1.4.2
kiwisolver==1.4.7
lazy_loader==0.4
matplotlib==3.9.3
networkx==3.4.2
numpy==2.2.0
packaging==24.2
pillow==11.0.0
pydicom==3.0.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
scikit-image==0.24.0
scikit-learn==1.6.0
scipy==1.14.1
six==1.17.0
threadpoolctl==3.5.0
tifffile==2024.9.20
vtk==9.4.0